refactor(queue): separate inspector and lane sections
